First approximately working version
- rake frank:init_features will create starter features
- rake frank:run will run features
- rake frank:symbiote will run the simulator with symbiote enabled

Note: frank:symbiote will leave behind the symbiote instrumentation
so that subsequent simulator runs will run symbiote. There's a bug
in how the build system cleans up resources that causes this. If
it doesn't get fixed I'll add something neater later.

I'm tracking the version of frank-cucumber with the versions of this,
so that e.g. 1.1.8.* correspond to frank-cucumber version 1.1.8